Tweaking

+++ Follower/LineFollower.h
@@ -25,7 +25,7 @@
 extern int followerState;
 extern bool isRealigning;
 
-extern int previousTime;
+extern unsigned long previousTime;
 extern float readLeft;
 extern float readRight;
 
@@ -44,7 +44,7 @@ MotorSpeeds driveMotorsBasic(float controller, float adjustedSpeed, float speedO
 MotorSpeeds driveMotorsPID(float controller, float derivative);
 
 float getLaneError();
-void followLaneAnalog(long currentTime);
+void followLaneAnalog(unsigned long currentTime);
 void determineStallPWM();//will determine the stallPWM of the robot with the current payload and battery power. It will add speed to the motors until the robot starts moving.
 
 #endif
